added admin

+++ src/main/java/com/example/boot_data_security/entities/User.java
@@ -18,8 +18,7 @@ import java.util.Set;
 public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-  //  private Long id;
+    private Long id;
     @Column(name = "username")
     private String username;
 
@@ -110,7 +109,7 @@ public class User implements UserDetails {
         this.record = record;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 }
